test: Load transaction signature v as uint64 (#1571)
A legacy EIP-155 transaction encodes `v` as chainId*2 + 35 + parity, so
`v` exceeds 0xff for any chainId >= 110 (e.g. chainId 300 gives
v=0x27c). The state-test loader narrowed `v` to `uint8_t` and
`Transaction::v` was `uint8_t`, so `from_json<uint8_t>` threw `value >
0xFF` and `evmone t8n` rejected such transactions while geth `evm t8n`
executes them.

Widen `Transaction::v` and its loader to `uint64_t`, matching the
`chainId` fix (#1570). `v` is only RLP-encoded for the transaction hash,
so the wider type is encoded correctly with no other behavior change.

Fixes #1487.

+++ test/utils/statetest_loader.cpp
@@ -456,7 +456,7 @@ state::Transaction from_json<state::Transaction>(const json::json& j)
     o.nonce = from_json<uint64_t>(j.at("nonce"));
     o.r = from_json<intx::uint256>(j.at("r"));
     o.s = from_json<intx::uint256>(j.at("s"));
-    o.v = from_json<uint8_t>(j.at("v"));
+    o.v = from_json<uint64_t>(j.at("v"));
 
     return o;
 }
